requirements.txt
Flask==2.3.3
PyJWT==2.8.0
python-dotenv==1.0.0
Werkzeug==2.3.7
tabulate==0.9.0 